Moved last file to TS, renamed TS files, fixes #52
