--- index_old.html
@@ -1,63 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <title>Gitdone Quick</title>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="https://www.w3schools.com/w3css/4/w3.css">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Raleway">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-    <link rel="stylesheet" href="css/bootstrap.css">    
-    <link rel="stylesheet" href="css/style.css">
-</head>
-
-<body class="w3-light-grey w3-content" style="max-width: 1600px">
-    <section class="container">
-        <form action="">
-            <div id="username">
-                <input type="text" placeholder="Username" value="DTasev" name="username" />
-            </div>
-            <div id="api-key">
-                <input type="password" placeholder="API Key" name="api-key" />
-                <a href="https://github.com/settings/tokens/new" target="_blank">Get API token</a>
-            </div>
-        </form>
-        <section>
-            <div id="repo-filter" class="inline-block">
-                <input type="text" placeholder="Filter repository names" />
-            </div>
-            <div id="error-message" class="inline-block"></div>
-        </section>
-        <div class="one pretty-links">
-            <table class="table-bordered">
-                <thead>
-                    <tr>
-                        <th>Repo name</th>
-                        <th>Ext</th>
-                    </tr>
-                </thead>
-                <tbody id="repository-list">
-
-                </tbody>
-            </table>
-        </div>
-
-        <div class="two pretty-links">
-            <table class="table-bordered">
-                <thead>
-                    <tr>
-                        <th>Issue name</th>
-                    </tr>
-                </thead>
-                <tbody id='issues-list'>
-                </tbody>
-            </table>
-        </div>
-    </section>
-    <script src="js/lib/jquery-3.2.1.js "></script>
-    <script src="js/lib/bootstrap.js "></script>
-    <script src="js/src/main.js "></script>
-</body>
-
-</html>
