+++ js/ts/load-icon.ts
